move getUnpickedPRs
